+++ src/lib/packages/forms/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @cluue/forms
 
+## 2.0.10
+
+### Patch Changes
+
+-   View Date in the Calendar and update gray color
+-   Updated dependencies
+    -   @cluue/icons@2.0.10
+    -   @cluue/base@2.0.10
+
+## 2.0.9
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @cluue/icons@2.0.9
+    -   @cluue/base@2.0.9
+
 ## 2.0.8
 
 ### Patch Changes